chore: refactor classes
